+++ premake5.lua
@@ -13,6 +13,7 @@ workspace "traffic3"
 		end
 	end
 	flags { "StaticRuntime", "MultiProcessorCompile" }
+  includedirs { "%{cfg.basedir}/src/main", "%{cfg.basedir}/3rdparty/easyloggingpp/src" }
 	filter "configurations:Debug"
     defines { "DEBUG" }
     flags { "Symbols" }
@@ -29,32 +30,34 @@ project "common"
 	kind "StaticLib"
 	language "C++"
   files { "src/main/lib/**.h", "src/main/lib/**.cpp" }
-	includedirs { "%{cfg.basedir}/3rdparty/easyloggingpp/src" }
+
+project "protocol"
+  kind "StaticLib"
+  language "C++"
+  files { "src/main/protocol/**.h", "src/main/protocol/**.cpp" }
+
+project "contentmanager"
+  kind "StaticLib"
+  language "C++"
+  files { "src/main/contentmanager/**.h", "src/main/contentmanager/**.cpp" }
 
 project "rest"
   kind "StaticLib"
   language "C++"
   files { "src/main/rest/**.h", "src/main/rest/**.cpp" }
-	includedirs { "%{cfg.basedir}/src/main/lib", "%{cfg.basedir}/3rdparty/easyloggingpp/src" }
 
 project "traffic3"
 	kind "ConsoleApp"
-	links { "common", "rest" }
+	links { "common", "rest", "contentmanager", "protocol" }
 	language "C++"
-  includedirs { "%{cfg.basedir}/src/main/lib", "%{cfg.basedir}/3rdparty/easyloggingpp/src",
+  includedirs { "%{cfg.basedir}/src/main", "%{cfg.basedir}/3rdparty/easyloggingpp/src",
     "%{cfg.basedir}/3rdparty/cmdline", "%{cfg.basedir}/src/main/rest" }
 	files { "src/main/console/**.h", "src/main/console/**.cpp" }
 
 project "traffic3_test"
 	kind "ConsoleApp"
-  links { "common" }
+  links { "common", "protocol", "rest", "contentmanager" }
 	language "C++"
-  includedirs { "%{cfg.basedir}/src/main/lib", "%{cfg.basedir}/3rdparty/Catch/single_include", "%{cfg.basedir}/3rdparty/hippomocks/HippoMocks", "%{cfg.basedir}/3rdparty/easyloggingpp/src" }
+  includedirs { "%{cfg.basedir}/src/main", "%{cfg.basedir}/3rdparty/Catch/single_include",
+    "%{cfg.basedir}/3rdparty/hippomocks/HippoMocks", "%{cfg.basedir}/3rdparty/easyloggingpp/src" }
 	files { "src/test/lib/**.h", "src/test/lib/**.cpp" }
-
-project "rest_test"
-  kind "ConsoleApp"
-  links { "common", "rest" }
-  language "C++"
-  includedirs { "%{cfg.basedir}/src/main/lib", "%{cfg.basedir}/src/main/rest", "%{cfg.basedir}/3rdparty/Catch/single_include", "%{cfg.basedir}/3rdparty/hippomocks/HippoMocks", "%{cfg.basedir}/3rdparty/easyloggingpp/src" }
-  files { "src/test/rest/**.h", "src/test/rest/**.cpp" }
