+++ src/main/contentmanager/contentmanagertype.h
