+++ src/main/protocol/protocol_udp.cpp
@@ -28,7 +28,7 @@ USA.
 #include <WS2tcpip.h>
 #endif
 #include <thread>
-#include "logging.h"
+#include "lib/logging.h"
 #include "protocol_udp.h"
 
 ProtocolUDP::ProtocolUDP() : Protocol() {
